fix: PR #137 리뷰 피드백 반영
- extract_branch_from_command: `--` 이후 파일 경로 무시 (git checkout -- file.txt)
- resolveWorkspaceId 중복 제거: workspace-utils.ts로 추출
- 미사용 OscCondition variant 제거 (MessageStartsWith, Not)
- Claude 태스크 전환 알림에 notifyGateArmed 게이팅 적용
